+++ Samples/macOS-Swift/macOS-Swift/AppDelegate.swift
@@ -12,7 +12,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             options.sessionTrackingIntervalMillis = 5_000
             // Sampling 100% - In Production you probably want to adjust this
             options.tracesSampleRate = 1.0
-            options.enableFileIOTracking = true
+            options.enableFileIOTracing = true
             if ProcessInfo.processInfo.arguments.contains("--io.sentry.profiling.enable") {
                 options.profilesSampleRate = 1
             }