+++ Sources/Sentry/Public/SentryUser.h
@@ -3,6 +3,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+NS_SWIFT_NAME(User)
 @interface SentryUser : NSObject <SentrySerializable, NSCopying>
 
 /**