+++ Sources/Sentry/include/SentryPerformanceTrackingIntegration.h
@@ -8,7 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
  * Integration to setup automatic performance tracking.
  *
  * Automatic UI performance setup can be avoided by setting
- * enableAutoPerformanceTracking to NO
+ * enableAutoPerformanceTracing to NO
  * in SentryOptions during SentrySDK initialization.
  */
 @interface SentryPerformanceTrackingIntegration : SentryBaseIntegration <SentryIntegrationProtocol>